+++ examples/1_planar_slicing_simple/example_1_planar_slicing_simple.py
@@ -12,7 +12,7 @@ from compas_slicer.post_processing import seams_smooth
 from compas_slicer.print_organization import PlanarPrintOrganizer
 from compas_slicer.print_organization import set_extruder_toggle
 from compas_slicer.print_organization import add_safety_printpoints
-from compas_slicer.print_organization import set_linear_velocity
+from compas_slicer.print_organization import set_linear_velocity_constant
 from compas_slicer.print_organization import set_blend_radius
 from compas_slicer.utilities import save_to_json
 from compas_viewers.objectviewer import ObjectViewer
@@ -101,7 +101,7 @@ def main():
     # ==========================================================================
     set_extruder_toggle(print_organizer, slicer)
     add_safety_printpoints(print_organizer, z_hop=10.0)
-    set_linear_velocity(print_organizer, "constant", v=25.0)
+    set_linear_velocity_constant(print_organizer, v=25.0)
     set_blend_radius(print_organizer, d_fillet=10.0)
 
     # ==========================================================================