+++ examples/2_curved_slicing/ex2_curved_slicing_advanced.py
@@ -10,6 +10,7 @@ from compas_slicer.pre_processing import create_mesh_boundary_attributes
 from compas_slicer.print_organization import InterpolationPrintOrganizer
 from compas_slicer.print_organization import set_extruder_toggle
 from compas_slicer.print_organization import add_safety_printpoints
+from compas_slicer.print_organization import smooth_printpoints_up_vectors
 from compas_slicer.post_processing import generate_brim
 import time
 
@@ -34,9 +35,7 @@ def main():
         'target_HIGH_smooth_union': [True, 7],  # boolean, blend_radius
         'avg_layer_height': 5.0,  # controls number of curves that will be generated
         'min_layer_height': 0.1,
-        'max_layer_height': 50.0,  # 2.0,
-        'layer_heights_smoothing': [False, 3, 0.5],  # boolean, iterations, strength
-        'up_vectors_smoothing': [False, 3, 0.5]  # boolean, iterations, strength
+        'max_layer_height': 50.0  # 2.0
     }
 
     ### --- Load initial_mesh
@@ -95,13 +94,14 @@ def main():
     # --- print organization
     if PRINT_ORGANIZER:
         filenames = utils.get_all_files_with_name('curved_slicer_', '.json', OUTPUT_PATH)
-        slicers = [BaseSlicer.from_data(utils.load_from_json(OUTPUT_PATH, filename)) for filename in filenames]
+        slicers = [InterpolationSlicer.from_data(utils.load_from_json(OUTPUT_PATH, filename)) for filename in filenames]
         for i, slicer in enumerate(slicers):
 
             print_organizer = InterpolationPrintOrganizer(slicer, parameters, DATA_PATH)
             print_organizer.create_printpoints()
             set_extruder_toggle(print_organizer, slicer)
             add_safety_printpoints(print_organizer, z_hop=10.0)
+            smooth_printpoints_up_vectors(print_organizer, strength=0.5, iterations=10)
 
             ### --- Save printpoints dictionary to json file
             printpoints_data = print_organizer.output_printpoints_dict()