+++ src/compas_slicer/print_organization/curved_print_organization/__init__.py
@@ -3,6 +3,5 @@ from __future__ import division
 from __future__ import print_function
 
 from .base_boundary import *  # noqa: F401 E402 F403
-from .vertical_connectivity import *  # noqa: F401 E402 F403
 
 __all__ = [name for name in dir() if not name.startswith('_')]