+++ src/compas_slicer/print_organization/planar_print_organizer.py
@@ -2,7 +2,9 @@ import logging
 from compas_slicer.print_organization import BasePrintOrganizer
 import compas_slicer.utilities as utils
 from compas_slicer.geometry import PrintPoint
+from compas.geometry import Vector
 import progressbar
+import compas_slicer
 
 logger = logging.getLogger('logger')
 
@@ -20,6 +22,7 @@ class PlanarPrintOrganizer(BasePrintOrganizer):
     """
 
     def __init__(self, slicer):
+        assert isinstance(slicer, compas_slicer.slicers.PlanarSlicer), 'Please provide a PlanarSlicer'
         BasePrintOrganizer.__init__(self, slicer)
 
     def __repr__(self):
@@ -31,6 +34,11 @@ class PlanarPrintOrganizer(BasePrintOrganizer):
         logger.info('Creating print points ...')
         with progressbar.ProgressBar(max_value=self.slicer.number_of_points) as bar:
 
+            # fast method for getting the closest mesh normals to all the printpoints
+            all_pts = [pt for layer in self.slicer.layers for path in layer.paths for pt in path.points]
+            closest_fks, projected_pts = utils.pull_pts_to_mesh_faces(self.slicer.mesh, all_pts)
+            normals = [Vector(*self.slicer.mesh.face_normal(fkey)) for fkey in closest_fks]
+
             for i, layer in enumerate(self.slicer.layers):
                 self.printpoints_dict['layer_%d' % i] = {}
 
@@ -38,18 +46,16 @@ class PlanarPrintOrganizer(BasePrintOrganizer):
                     self.printpoints_dict['layer_%d' % i]['path_%d' % j] = []
 
                     for k, point in enumerate(path.points):
-                        normal = utils.get_normal_of_path_on_xy_plane(k, point, path, self.slicer.mesh)
 
-                        printpoint = PrintPoint(pt=point, layer_height=self.slicer.layer_height, mesh_normal=normal)
+                        n = normals[count]
+                        if layer.is_raft or layer.is_brim:  # then project normal on the xy plane
+                            n = Vector(n[0], n[1], 0.0)
+
+                        printpoint = PrintPoint(pt=point, layer_height=self.slicer.layer_height, mesh_normal=n)
 
                         self.printpoints_dict['layer_%d' % i]['path_%d' % j].append(printpoint)
                         bar.update(count)
                         count += 1
-
-    def check_printpoints_feasibility(self):
-        """ Check the feasibility of the print points """
-        # TODO
-        raise NotImplementedError
 
 
 if __name__ == "__main__":